requirements.txt
streamlit
transformers
torch
pillow
openai
langchain
langchain-community